Move xds server to container directory
